Update directory architecture
