+++ CITA_Monitor_Agent_Docker/DockerCompose_Files/.env
@@ -1,5 +1,5 @@
 # container hostname
-hostname=EP_Host_01
+hostname=EP_Host_192_168_2_174
 
 # prometheus_node_exporter outside port and inside port
 node_out_port=9100
@@ -20,3 +20,10 @@ agent_in_port=1920
 # CITA node target ip and port
 target_ip=192.168.2.174
 target_port=1337
+
+# CITA node directory analysis
+# Support multiple directories, split by ',' 
+node_dir=/github/cita_secp256k1_sha3/monitor_chain_02/0
+
+# CITA node id
+node_id=0
