+++ CITA_Monitor_Agent_Docker/DockerImage_Build_Files/Dockerfile
@@ -1,5 +1,5 @@
 ## 一个基础的 python 运行环境
-FROM python:3
+FROM python
 ## 设置工作目录
 WORKDIR /config
 ## 将当前系统文件夹内容复制到容器的 app 目录
@@ -8,5 +8,7 @@ ADD . /config
 RUN pip3 install -r requirements.txt
 RUN mv cita-cli /bin/cita-cli && chmod +x /bin/cita-cli
 ENV OpenPort=1920
+ENV Dir='/tmp'
+ENV NodeID=0
 ## 运行命令
-CMD "python3" "cita_agent_by_cita-cli.py" "$Node" "$OpenPort"
+CMD "python3" "cita_agent_by_cita-cli.py" "$Node" "$OpenPort" "$Dir" "$NodeID"
